refactor(mapStyleDuck): rename
