+++ src/base/static/components/organisms/offline-download-menu.js
@@ -12,7 +12,7 @@ import {
   modalStyles,
 } from "../atoms/layout";
 import { offlineConfigPropType } from "../../state/ducks/map-config";
-import { mapSourcesPropType, mapSourcesSelector } from "../../state/ducks/map";
+import { mapSourcesPropType, mapSourcesSelector } from "../../state/ducks/map-style";
 
 import Modal from "react-modal";
 import { connect } from "react-redux";
