+++ src/base/static/state/ducks/map-style.d.ts
