+++ src/base/static/state/ducks/map-style.js
